Move ProcessingScreen to common components folder
